style: apply make fix (blank lines + drop redundant casts)
Formatting-only follow-up from make fix on the versioning managers/uploader:
cs-fixer blank lines + rector removal of redundant (int)/(string) casts
(getSize/setSize/getPath are already non-null). No behaviour change; suite
green (2813).

+++ src/Module/Ged/Document/Service/GedDocumentUploader.php
@@ -118,7 +118,7 @@ final readonly class GedDocumentUploader
         return [
             'filePath' => $relativePath,
             'fileName' => $newFilename,
-            'size' => (int) (@filesize(Path::join($this->uploadDir, $relativePath)) ?: 0),
+            'size' => @filesize(Path::join($this->uploadDir, $relativePath)) ?: 0,
             'width' => $dimensions[0],
             'height' => $dimensions[1],
         ];
